updated code to resolve warnings
separated out scss into two files

refactoring profile object

moved profile export to separate file

refactored code to remove eslint errors

removed extra line

renamed profile.js to types.js

renamed import to changed named
