+++ client/app/components/Header/HeaderProfile.jsx
@@ -2,11 +2,10 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBell } from '@fortawesome/free-solid-svg-icons';
-import { Avatar } from '../Avatar';
-import type { Profile } from './index';
+import type { Profile } from './types';
 import { Notifications } from '../../widgets/Notifications';
-// eslint-disable-next-line import/no-cycle
-import css from './Header.scss';
+import { Avatar } from '../Avatar';
+import css from './HeaderProfile.scss';
 import globalCSS from '../../styles/_global.scss';
 
 export type Props = {
